Suppress default log appender layout warning
The RocketmqLog4j2Appender will always initialize a SerializedLayout in its constructor, which will cause a warning of 'SerializedLayout is deprecated due to the inherent security weakness in Java Serialization'.

In order to suppress this warning, I moved it to the appender build method, so that the SerializedLayout will be initialized only when the layout is not set.