fixed java.lang.IllegalStateException: Shutdown in progress on app shutdown before started
in k8s cluster, when deploy a workload,  in pod A  the app is starting. then deploy again right now, pod A will be destoryed. sometimes error log: java.lang.IllegalStateException: Shutdown in progress

In Idea, we break a point on this line, then start the app. then click shutdown, will reproduce this.